itinaretion 2 complete

+++ rooms-app/routes/room.routes.js
@@ -9,9 +9,11 @@ const Room = require('../models/Room.model');
 router.get('/', (req, res, next) => {
   // Iteration #2: List the room
   Room.find()
-      .then(rooms => res.render('rooms/list', { rooms }))
+      .then(rooms => res.render('rooms/list', { rooms , user:req.session.currentUser  }))
       .catch(err => console.log(err))
 });
+
+
 
 router.get('/create', isLoggedIn, (req, res, next) => {
   // Add a new room
@@ -35,45 +37,53 @@ router.post('/create', (req, res, next) => {
   .catch(err => console.log(err))
 });
 
-router.get('/:id/edit', (req, res, next) => {
+router.get('/:id/edit', async (req, res, next) => {
   // Iteration #4: Update the rooms
 
   console.log("req.params",req.params)
   const { id } = req.params;
 
-
-  Room.findById(id)
+  const theRoom =await Room.findById(id)
+  if(theRoom.owner.toString() === req.session.currentUser._id){
+    Room.findById(id)
     .then(foundRoom => res.render('rooms/update-form', foundRoom))
     .catch(err => console.log(err))
-  
+  }else {
+    res.redirect("/rooms" )
+  }
+ 
 });
 
-router.post('/:id/edit', (req, res, next) => {
+router.post('/:id/edit', async (req, res, next) => {
   // Iteration #4: Update the rooms
   const { name, description, imageUrl } = req.body;
   const { id } = req.params;
 
-  Room.findByIdAndUpdate(id, {name, description, imageUrl})
+  const theRoom = await Room.findById(id)
+  if(theRoom.owner.toString() === req.session.currentUser._id){
+      Room.findByIdAndUpdate(id, {name, description, imageUrl})
       .then(() => res.redirect('/rooms'))
       .catch(err => console.log(err))
+    } 
   
 });
 
-router.post('/:id/delete', (req, res, next) => {
+router.post('/:id/delete', async (req, res, next) => {
   // Iteration #5: Delete the room
   const { id } = req.params;
 
-  Room.findByIdAndDelete(id)
-      .then(() => res.redirect('/rooms'))
-      .catch(err => console.log(err))
+  const theRoom = await Room.findById(id)
+  
+  if(theRoom.owner.toString() === req.session.currentUser._id){
+    Room.findByIdAndDelete(id)
+    .then(() => res.redirect('/rooms'))
+    .catch(err => console.log(err)) 
+  } else {
+    res.redirect("/rooms" )
+  }
+
+
+
 });
 
-// router.get('/list', (req, res) => {
-//   Room.find()
-//       .populate('owner')
-//       .then((roomsList) => {
-//           res.render('room/list', { roomsList })
-//       })
-//       .catch((err) => console.error(`Error while listing movies: ${err}`))
-// })
 module.exports = router;
